Move authentication and macaroon to modules
