+++ modules/macaroon.py
